refactor: move incoming call logic to separate package
